% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{calibration_likelihood}
\alias{calibration_likelihood}
\title{Specifics the model likelihood}
\usage{
calibration_likelihood(b0, fixed_parameters, ..., data)
}
\arguments{
\item{b0}{Vector of parameter names that we want to fit.}

\item{fixed_parameters}{List of parameters that are fixed.}

\item{...}{
  Arguments passed on to \code{\link[=run_model]{run_model}}
  \describe{
    \item{\code{max_t}}{Simulation maximum time. Default: 3650 days.}
    \item{\code{cohort_step_size}}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}
    \item{\code{dt}}{Time steps to run the model over. Default: 0.25 days.}
    \item{\code{init_conds}}{Initial conditions to run the model. List. Default: \code{NULL}. If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection.
All other people are assumed to be susceptible to their primary infection.}
  }}
}
\value{
Log-likelihood.
}
\description{
Function to run the model and calculate the likelihood.
}
