% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{constrained_max_likelihood}
\alias{constrained_max_likelihood}
\title{Constrained maximum-likelihood estimation of the b0, b1 and phi parameters from incidence data.}
\usage{
constrained_max_likelihood(
  fixed_parameter_list,
  data,
  scale_parameters,
  cohort_step_size,
  dt
)
}
\arguments{
\item{fixed_parameter_list}{List of parameters that are fixed.}

\item{data}{List of data frames of incidence data. For all data frames "incidence", "time" and "age_chr" columns must be included.
For incidence data frames a "value" column is needed to contain the incidence values.
For the prevalence data frame a "sample" column is required with the number of people tested and a "positive" column with the number of people that test positive.}

\item{scale_parameters}{List of lower and upper boundaries on the fitted parameters.
Default: \code{NULL}. If \code{NULL} the parameters are not transformed.
If a list is provided the parameters log-likelihood is calculated assuming the \code{fitted_parameters}
have already been scaled between 0 and 1 with the given lower and upper limits.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{dt}{Time steps to run the model over. Default: 0.25 days.}
}
\value{
List of fitted parameter values for b0, b1 and phi.
}
\description{
Helper function to assess the times and ages in the data, and return suitable parameters to run the model. Initial parameter values are set to 0.
}
