% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim-package.R
\docType{package}
\name{RSVsim-package}
\alias{RSVsim-package}
\title{RSVsim: RSV Transmission Model}
\description{
Deterministic, ODE model of Respiratory Synctial Virus (RSV) transmission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IsaacStopard/RSVsim}
  \item \url{https://isaacstopard.github.io/RSVsim/}
  \item Report bugs at \url{https://github.com/IsaacStopard/RSVsim/issues}
}

}
\author{
\strong{Maintainer}: Isaac Stopard \email{isaacstopard@hotmail.co.uk} (\href{https://orcid.org/0000-0003-4612-5004}{ORCID})

Authors:
\itemize{
  \item Alexandra Hogan \email{alexandra.hogan@unsw.edu.au}
}

}
\keyword{internal}
