% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim-package.R
\name{RSVsim}
\alias{RSVsim}
\title{RSVsim: A package for simulating Respiratory Synctial Virus (RSV) transmission.}
\description{
RSVsim: A package for simulating Respiratory Synctial Virus (RSV) transmission.
}
