% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_ABC_rejection}
\alias{RSVsim_ABC_rejection}
\title{Function to run an Approximate Bayesian Computation (ABC) rejection algorithm}
\usage{
RSVsim_ABC_rejection(
  target,
  epsilon,
  summary_fun,
  dist_fun,
  prior_fun,
  n_prior_attempts,
  nparticles,
  used_seeds_all,
  ncores = 1,
  fitted_parameter_names,
  fixed_parameter_list,
  times = seq(0, 365.25 * 5, 0.25),
  cohort_step_size = 1/12 * 365.25,
  warm_up = 365.25 * 4
)
}
\arguments{
\item{target}{Values to fit to.}

\item{epsilon}{Acceptable error for each target.}

\item{summary_fun}{Function to calculate the summary statistics equivalent to the target values.}

\item{dist_fun}{Function to the calculate the error between the target and \code{summary_fun} outputs.}

\item{prior_fun}{Function to sample from the priors for all parameters. Must return a vector.}

\item{n_prior_attempts}{Number of random samples from the prior to attempt for each accepted particle.}

\item{nparticles}{Integer. Number of samples from the approximate posterior.}

\item{used_seeds_all}{Vector. Seeds used when generating the prior samples for each accepted particle.}

\item{ncores}{Number of cores. If greater than one then it is run in parallel.}

\item{fitted_parameter_names}{Vector of names of the parameters that are being estimated. To fit vectors or matrices the individual element must be identified in the fitted_parameter_names.}

\item{fixed_parameter_list}{List of the original parameter values to run the model.}

\item{times}{Simulation times. Default: 0 to 365.25 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 1 month. If \code{is.numeric(cohort_step_size) == FALSE} then cohort ageing is not applied. Can have a maximum of 3 decimal places.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
List of fixed parameters, max_t and warm_up.
}
\description{
Runs the ABC-rejection algorithm.
}
