% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_ABC_SMC}
\alias{RSVsim_ABC_SMC}
\title{Function to run a Approximate Bayesian Computation Sequential Monte Carlo (ABC-SMC) algorithm}
\usage{
RSVsim_ABC_SMC(
  target,
  epsilon_matrix,
  summary_fun,
  dist_fun,
  prior_fun,
  prior_dens_fun,
  particle_low,
  particle_up,
  nparticles,
  ncores = 1,
  fitted_parameter_names,
  fixed_parameter_list,
  times = seq(0, 365 * 5, 0.25),
  cohort_step_size = 0.2 * 365,
  init_conds = NULL,
  warm_up = 365 * 4
)
}
\arguments{
\item{target}{Values to fit to.}

\item{epsilon_matrix}{Matrix of acceptable error for each target (columns) for each generation (rows). The row number of is used to determine the number of generations.}

\item{summary_fun}{Function to calculate the summary statistics equivalent to the target values. Should take one argument: the model outputs.}

\item{dist_fun}{Function to the calculate the error between the target and \code{summary_fun} outputs. Should take one argument: the model outputs.}

\item{prior_fun}{Function to sample from the priors for all parameters. Must return a vector.}

\item{prior_dens_fun}{Function that calculates the joint probability density of all parameters given the prior distributions.}

\item{nparticles}{Number of samples from the approximate posterior for each generation.}

\item{ncores}{Number of cores. If greater than one then it is run in parallel.}

\item{fitted_parameter_names}{Vector of names of the parameters that are being estimated.}

\item{fixed_parameter_list}{List of parameter values to run the model excluding the fitted parameters.}

\item{times}{Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{init_conds}{Initial conditions to run the model - these are given as prevalence and the initial conditions calculated in this function. List. Default: \code{NULL}.
If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection, which is seeded at the beginning of the simulation.
All other people are assumed to be susceptible to their primary infection.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
List of fixed parameters, max_t and warm_up.
}
\description{
----- NOT FINISHED ------ implementation of an ABC-SMC algorithm.
}
