% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_abs_dist_fun}
\alias{RSVsim_abs_dist_fun}
\title{Distance function to calculate the absolute difference between two vectors.}
\usage{
RSVsim_abs_dist_fun(target, target_star)
}
\arguments{
\item{target}{First vector of metrics.}

\item{target_star}{Second vector of metrics.}
}
\value{
Absolute difference.
}
\description{
Calculates the absolute distance. Vectorised.
}
