% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_amplitude}
\alias{RSVsim_amplitude}
\title{Summary function to calculate the age-specific amplitude of changes in incidence over the simulation.}
\usage{
RSVsim_amplitude(out)
}
\arguments{
\item{out}{\code{RSVsim_run_model} function output.}
}
\value{
Summary statistics (amplitude by age-group) from the simulation output (dataframe).
}
\description{
Function calculates the difference between maximum and minimum incidence stratified by age over the whole simulation (excluding warm up).
}
