% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{RSVsim_calibration_parameters}
\alias{RSVsim_calibration_parameters}
\title{Get parameters for the RSVsim_log_likelihood function from the data}
\usage{
RSVsim_calibration_parameters(
  fitted_parameter_names = c("b0", "b1", "phi"),
  data,
  overrides = list(),
  country = "United Kingdom",
  cohort_step_size = 10,
  warm_up = NULL
)
}
\arguments{
\item{fitted_parameter_names}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}

\item{data}{Data frame of incidence data: "incidence", "time" and "age_chr" columns must be included.}

\item{overrides}{List of default parameters to change.}

\item{country}{Country for use in the \code{contact_matrix} function in the \code{socialmixr} package. Can be given as country name or 2 digit ISO code. United Kingdom default.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
List of fixed parameters, max_t and warm_up.
}
\description{
Helper function to assess the times and ages in the data, and return a list of suitable fixed parameters and
the model control option values given as arguments to the function.
}
