% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_contact_matrix.R
\name{RSVsim_contact_matrix}
\alias{RSVsim_contact_matrix}
\title{Create contact matrix for RSVsim}
\usage{
RSVsim_contact_matrix(
  country = "United Kingdom",
  age.limits = c(seq(0, 5, 1/12), seq(10, 70, 5))
)
}
\arguments{
\item{country}{Country for use in the \code{contact_matrix} function in the \code{socialmixr} package. Can be given as country name or 2 digit ISO code. United Kingdom default.}

\item{age.limits}{Lower limits of the age groups to run the simulation. Ages must be in years.}
}
\value{
List including the contact matrix (\code{matrix_mean} is the mean per person, \code{matrix_contacts} is the total contacts),
the age limits, the age distribution used to calculate the total contacts and the maximum age in the contact data.
}
\description{
Uses the \code{socialmixr} package to access the POLYMOD data and calculate per-person mean contact matrix.
If age groupings smaller than 1-year are required, the function calculates the total contacts in the
population with the nearest integer age groupings (required for \code{socialmixr} package) and splits the total contacts
into smaller age groupings by dividing them uniformly.
}
