% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_fixed_parameters}
\alias{RSVsim_fixed_parameters}
\title{Get fixed parameters for the RSVsim_calibrate function from the data}
\usage{
RSVsim_fixed_parameters(
  fitted_parameter_names = c("b0"),
  data,
  overrides = list(),
  country = "United Kingdom"
)
}
\arguments{
\item{fitted_parameter_names}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}

\item{data}{Dataframe of data to fit to: "age_chr" columns must be included.}

\item{overrides}{List of default parameters to change.}

\item{country}{Country for use in the \code{contact_matrix} function in the \code{socialmixr} package. Can be given as country name or 2 digit ISO code. United Kingdom default.}
}
\value{
List of fixed parameters, max_t and warm_up.
}
\description{
Helper function to assess the ages in the data, and return a list of suitable fixed parameters and
the model control option values given as arguments to the function.
}
