% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{RSVsim_max_likelihood_dust}
\alias{RSVsim_max_likelihood_dust}
\title{Constrained maximum-likelihood estimation of the b0, b1 and phi parameters from incidence data.}
\usage{
RSVsim_max_likelihood_dust(
  data,
  fitted_parameter_names,
  fixed_parameters,
  times,
  cohort_step_size = 10,
  init_conds = NULL,
  warm_up = NULL,
  scale_parameters = FALSE,
  lower_ll,
  upper_ll
)
}
\arguments{
\item{fitted_parameter_names}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}

\item{times}{Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{init_conds}{Initial conditions to run the model. List. Default: \code{NULL}. If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection.
All other people are assumed to be susceptible to their primary infection.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}

\item{scale_parameters}{Boolean operator indicating whether to scale the parameters between 0 and 1 to help fitting multiple parameters on difference scales. Default: \code{FALSE}.}

\item{lower_ll}{Vector of lower and boundaries on the fitted parameters.}

\item{upper_ll}{Vector of upper boundaries on the fitted parameters.}
}
\value{
List of fitted parameter values.
}
\description{
Wrapper function for the \code{stats::nlminb function} (non-linear box-constrained quasi-Newton optimisation).
Used to maximise the \code{RSVsim_log_likelihood_dust} function with data and parameters formatted by the \code{RSVsim_calibration_parameters} function.
}
