% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{RSVsim_MCMC}
\alias{RSVsim_MCMC}
\title{Metropolis-Hastings Markov-chain Monte Carlo estimation of the \code{b0}, \code{b1} and \code{phi} parameters from incidence data.}
\usage{
RSVsim_MCMC(
  data,
  cores,
  nchains,
  mcmc_iter,
  mcmc_warm_up,
  prior_functions_list,
  initial_mat,
  fitted_parameter_names,
  fixed_parameters,
  times,
  cohort_step_size,
  init_conds,
  warm_up,
  lower_ll,
  upper_ll
)
}
\arguments{
\item{fitted_parameter_names}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}

\item{times}{Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{init_conds}{Initial conditions to run the model. List. Default: \code{NULL}. If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection.
All other people are assumed to be susceptible to their primary infection.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
MCMC samples for b0, b1 and phi.
}
\description{
Wrapper function for the fits the b0, b1 and and phi parameters.
}
