% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_peak}
\alias{RSVsim_peak}
\title{Summary function to calculate the age-specific timing of peak incidence.}
\usage{
RSVsim_peak(out)
}
\arguments{
\item{out}{\code{RSVsim_run_model} function output.}
}
\value{
Summary statistics (peak incidence time ordered by age-group) from the simulation output (dataframe).
}
\description{
Function calculates the time when incidence peaks stratified by age over the whole simulation (excluding warm up).
}
