% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_hospitalisation_functions.R
\name{RSVsim_prob_hosp_given_inf}
\alias{RSVsim_prob_hosp_given_inf}
\title{Probability of hospitalisation given unvaccinated infection}
\usage{
RSVsim_prob_hosp_given_inf(
  hosp_rate,
  age_limits_hosp_rate,
  hosp_rate_min_time,
  hosp_rate_max_time,
  sim_status_quo,
  VE_hosp_given_inf,
  VE_hosp_vacc_states
)
}
\arguments{
\item{hosp_rate}{Vector of age-specific RSV hospitalisation rates (per person) in the status quo scenario.}

\item{age_limits_hosp_rate}{Vector of ages corresponding to the first age of each age category in (1) the hospitalisation rates (\code{hosp_rate}) and (2) the probability of hospitalisation given infection (\code{prob_hosp_given_inf}). These ages must be present in the age_limits used to run the simulations.}

\item{hosp_rate_min_time}{Minimum simulation time to select when summing the hospitalisations to calculate the hospitalisation rates.}

\item{hosp_rate_max_time}{Maximum simulation time to select when summing the hospitalisations to calculate the hospitalisation rates.}

\item{sim_status_quo}{\code{RSVsim_run_model} output used to estimate the number of infections when calculating the probability of hospitalisation given infection.}

\item{VE_hosp_given_inf}{Matrix of age-specific vaccine efficacy against hospitalisation given infection (\code{VE_hosp_given_inf}). Rows must correspond to the ages in \code{age_limits}, and columns must correspond to the vaccinated states used in \code{RSVsim_run_model}.}

\item{VE_hosp_vacc_states}{vector of integers indicating the vaccinated states used in \code{RSVsim_run_model}.}
}
\value{
List of probability of hospitalisation given infection  and \code{sim_status_quo} model outputs with hospitalisations.
}
\description{
Calculates the age-specific probability of hospitalisation given infection, using observed hospitalisation rates per person among the population as a whole
and RSVsim to estimate number of infections per person.
}
