% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_run_model.R
\name{RSVsim_run_model}
\alias{RSVsim_run_model}
\title{Runs the RSV model using odin and dust}
\usage{
RSVsim_run_model(
  parameters,
  times = seq(0, 365.25 * 1, 0.25),
  cohort_step_size = 1/12 * 365.25,
  warm_up = NULL
)
}
\arguments{
\item{parameters}{List of parameters from \code{RSVsim_parameters} function.}

\item{times}{Simulation times. Default: 0 to 365.25 days with intervals of 0.25 days. The minimum interval size is 1E-5.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 1 month. If \code{is.numeric(cohort_step_size) == FALSE} then cohort ageing is not applied. Can have a maximum of 7 decimal places.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
Simulation output (dataframe). In the dataframe, age refers to the lowest age in the age group.
}
\description{
Function to run the transmission model with cohort aging. Prevalence, incidence between the given time-steps and the incidence rate per day are also calculated.
}
