% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_run_model_dust.R
\name{RSVsim_run_model_dust}
\alias{RSVsim_run_model_dust}
\title{Runs the RSV model using odin and dust}
\usage{
RSVsim_run_model_dust(
  parameters,
  times = seq(0, 3650, 0.25),
  cohort_step_size = 10,
  init_conds = NULL,
  warm_up = NULL
)
}
\arguments{
\item{parameters}{List of parameters from \code{get_params} function.}

\item{times}{Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{init_conds}{Initial conditions to run the model. List. Default: \code{NULL}. If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection.
All other people are assumed to be susceptible to their primary infection.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
Simulation output (dataframe). In the dataframe, age refers to the lowest age in the age group.
}
\description{
Function to run the transmission model with cohort aging.
}
