% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_shortest_periodic_dist_fun}
\alias{RSVsim_shortest_periodic_dist_fun}
\title{Distance function to calculate the shortest periodic difference between two vectors.}
\usage{
RSVsim_shortest_periodic_dist_fun(target, target_star, period)
}
\arguments{
\item{target}{First vector of metrics.}

\item{target_star}{Second vector of metrics.}

\item{period}{Cosine wave period.}
}
\value{
Absolute difference.
}
\description{
Calculates the absolute distance between two times when the times are circular. Vectorised.
}
