% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_total_incidence}
\alias{RSVsim_total_incidence}
\title{Summary function to calculate the age-specific total incidence over the simulation.}
\usage{
RSVsim_total_incidence(out)
}
\arguments{
\item{out}{\code{RSVsim_run_model} function output.}
}
\value{
Summary statistics (total incidence ordered by age-group) from the simulation output (dataframe).
}
\description{
Function calculates the total incidence stratified by age over the whole simulation (excluding warm up).
}
