% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{RSVsim_update_parameters}
\alias{RSVsim_update_parameters}
\title{Helper function to update the parameter list by names}
\usage{
RSVsim_update_parameters(
  fixed_parameter_list,
  fitted_parameter_names,
  fitted_parameter_values
)
}
\arguments{
\item{fixed_parameter_list}{List of parameters, should be the output of the \code{RSVsim_parameters} function.}

\item{fitted_parameter_names}{Vector of parameter names to be updated.}

\item{fitted_parameter_values}{Vector of updated parameter values.}
}
\value{
Parameter list.
}
\description{
Changes the values in the list of parameters to run the model.
To fit parameters stored within vectors or matrices, \code{fitted_parameter_names} can also include indexing such as "parameter_name[i]".
}
