% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_hospitalisation_functions.R
\name{RSVsim_VE_hosp_given_inf}
\alias{RSVsim_VE_hosp_given_inf}
\title{Vaccine efficacy against hospitalisation conditional on infection}
\usage{
RSVsim_VE_hosp_given_inf(VE_hosp, VE_inf)
}
\arguments{
\item{VE_hosp}{Vaccine efficacy against hospitalisation.}

\item{VE_inf}{Vaccine efficacy against infection.}
}
\value{
Vaccine efficacy against hospitalisation given infection (\code{VE_hosp_given_inf}).
}
\description{
Calculates the vaccine efficacy against hospitalisation given infection using the
overall vaccine efficacy against hospitalisation and vaccine efficacy against infection.
If the number ages in VE_inf are greater than the number of ages in VE_hosp we just select the
}
