% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs the RSV model}
\usage{
run_model(
  parameters,
  max_t = 2000,
  dt = 0.25,
  init_conds = NULL,
  save_final_states = FALSE
)
}
\arguments{
\item{parameters}{List of parameters from \code{get_params} function.}

\item{max_t}{Simulation maximum time. Default: 2000 days.}

\item{dt}{Time steps to run the model over. Default: 0.25 days.}

\item{init_conds}{Initial conditions to run the model. List. Default: \code{NULL}. If \code{NULL} 1\% RSV prevalence is assumed for people during the primary infection.
All other people are assumed to be susceptible to their primary infection.}

\item{save_final_states}{Boolean. Choose whether to save final model state as initial conditions for next simulation. Default: \code{FALSE}.}
}
\value{
Simulation output
}
\description{
Function to run the transmission model with cohort aging.
}
