% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.r
\name{while_fun}
\alias{while_fun}
\title{Function to return a single particle for the ABC-rejection algorithm.}
\usage{
while_fun(
  particle,
  target,
  epsilon,
  summary_fun,
  dist_fun,
  prior_fun,
  n_prior_attempts,
  used_seeds_all,
  fitted_parameter_names,
  fixed_parameter_list,
  times,
  cohort_step_size,
  init_conds,
  warm_up
)
}
\arguments{
\item{particle}{Particle number.}

\item{target}{Values to fit to.}

\item{epsilon}{Acceptable error for each target.}

\item{summary_fun}{Function to calculate the summary statistics equivalent to the target values.}

\item{dist_fun}{Function to the calculate the error between the target and \code{summary_fun} outputs.}

\item{prior_fun}{Function to sample from the priors for all parameters. Must return a vector.}

\item{n_prior_attempts}{Number of random samples from the prior to attempt for each accepted particle.}

\item{used_seeds_all}{Vector. Seeds used when generating the prior samples for each accepted particle.}

\item{fitted_parameter_names}{Vector of names of the parameters that are being estimated.}

\item{fixed_parameter_list}{List of parameter values to run the model excluding the fitted parameters.}

\item{times}{Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.}

\item{cohort_step_size}{Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.}

\item{init_conds}{Initial conditions to run the model - these are given as prevalence and the initial conditions calculated in this function. List. Default: \code{NULL}.
If \code{NULL}: 1\% RSV prevalence is assumed for people during the primary infection, which is seeded at the beginning of the simulation.
All other people are assumed to be susceptible to their primary infection.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\description{
Function to return a single particle for the ABC-rejection algorithm.
}
