% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters for RSVsim}
\usage{
get_parameters(overrides = list(), contact_population_list)
}
\arguments{
\item{overrides}{List of default parameters to change.}

\item{contact_population_list}{List of outputs from the \code{create_contact_matrix} function.}
}
\value{
Parameter list.
}
\description{
Provides the default parameters to run the RSV model.
}
