% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{RSVsim_calibration_parameters}
\alias{RSVsim_calibration_parameters}
\title{Get parameters for the RSVsim_log_likelihood function from the data}
\usage{
RSVsim_calibration_parameters(
  data,
  data_populations,
  country = "United Kingdom",
  warm_up = NULL
)
}
\arguments{
\item{data}{List of data frames of incidence data. For all data frames "incidence", "time" and "age_chr" columns must be included.
For incidence data frames a "value" column is needed to contain the incidence values.
For the prevalence data frame a "sample" column is required with the number of people tested and a "positive" column with the number of people that test positive.}

\item{data_populations}{Vector of total population sizes for each dataframe in \code{data}.}

\item{country}{Country for use in the \code{contact_matrix} function in the \code{socialmixr} package. Can be given as country name or 2 digit ISO code. United Kingdom default.}

\item{warm_up}{Length of time-points to exclude before calculating the likelihood. Default: \code{NULL}.}
}
\value{
List of fixed parameters, max_t and warm_up.
}
\description{
Helper function to assess the times and ages in the data, and return suitable parameters to run the model.
}
