% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_hospitalisation_functions.R
\name{RSVsim_hospitalisations}
\alias{RSVsim_hospitalisations}
\title{Calculate number of hospitalisations for \code{RSVsim_run_model} output}
\usage{
RSVsim_hospitalisations(
  prob_hosp_given_inf,
  age_limits_hosp_rate,
  sim_scenario,
  VE_HOSP,
  VE_HOSP_vacc_states
)
}
\arguments{
\item{prob_hosp_given_inf}{Age-specific probability of hospitalisation given unvaccinated infection. Ages must correspond to those in \code{age_limits_hosp_rate}.}

\item{age_limits_hosp_rate}{Vector of ages corresponding to the first age of each age category for the hospitalisation rates and probability of hospitalisation given infection. These ages must be present in the age_limits used to run the simulations.}

\item{sim_scenario}{\code{RSVsim_run_model} output. Hospitalisations are calculated for this scenario using the estimated probability of hospitalisation given infection.}

\item{VE_HOSP}{Matrix of age-specific vaccine efficacy against hospitalisation. Rows must correspond to the ages in \code{age_limits_hosp_rate}, and columns must correspond to the vaccinated states used in \code{RSVsim_run_model}.}

\item{VE_HOSP_vacc_states}{vector of integers indicating the vaccinated states used in \code{RSVsim_run_model}.}
}
\value{
sim_scenario dataframe with hospitalisations column
}
\description{
Function to estimate the number of hospitalisations from the modelled incidence and estimated probability of hospitalisation given unvaccinated infection.
}
