% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_parameters.R
\name{RSVsim_parameters}
\alias{RSVsim_parameters}
\title{Get parameters for RSVsim}
\usage{
RSVsim_parameters(overrides = list(), contact_population_list, fitted = NULL)
}
\arguments{
\item{overrides}{List of default parameters to change.}

\item{contact_population_list}{List of outputs from the \code{create_contact_matrix} function.}

\item{fitted}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}
}
\value{
Parameter list.
}
\description{
Provides the default parameters to run the RSV model.
}
