% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVsim_parameters.R
\name{RSVsim_parameters}
\alias{RSVsim_parameters}
\title{Get parameters for RSVsim}
\usage{
RSVsim_parameters(
  overrides = list(),
  contact_population_list,
  fitted_parameter_names = NULL
)
}
\arguments{
\item{overrides}{List of default parameters to change.}

\item{contact_population_list}{List of outputs from the \code{create_contact_matrix} function.}

\item{fitted_parameter_names}{Vector of parameter names to exclude from the list because they are fitted. Default: \code{NULL}.}
}
\value{
Parameter list. Default values: \cr
\code{b0}: 0.08. Transmission rate coefficient. \cr
\code{b1}: 0. Amplitude of seasonal forcing in the transmission rate. \cr
\code{phi}: 0. Phase shift of the seasonal forcing. \cr
\code{delta}: 1/4. Inverse of the latent period. \cr
\code{gamma_s}: 1/8. Inverse of the infectious period of subsequent (secondary) infections. \cr
\code{gamma_p}: 1/9. Inverse of the infectious period of first (primary) infections. \cr
\code{nu}: 1/200. Inverse of the duration of natural immunity. \cr
\code{omega_vect}: 0.35 if greater than 5 years old else 1. Reduced infectiousness due to immunity in early life. \cr
\code{prop_detected_vect}: age-specific proportion detected. \cr
\code{sigma_vect}: 0.7 if 1 months old, 0.8 if 2 months old, 0.9 if 3 months old, else 1. Reduced susceptibility due to immunity in early life. \cr
\code{alpha_vect} = 0.4 if less than 10 years old else 0.3. Reduced susceptibility to secondary infections in each age group. \cr
\code{nAges}: must be obtained from \code{contact_population_list}. Number of age groups - same length as age.limits. \cr
\code{age.limits}: must be obtained from \code{contact_population_list}. Lower age limits in years. \cr
\code{matrix_per_person}: must be obtained from \code{contact_population_list}. Contact matrix of mean contacts per person. \cr
\code{max_age}: must be obtained from \code{contact_population_list}. Maximum age in years. \cr
\code{age_chr}: Age limits as a character vector. \cr
\code{size_cohorts}: the differences in ages in days - used to calculate the size of the cohorts. \cr
\code{rel_sizes}: the relative size of each cohort (age category duration in days divided by the total number of days). \cr
\code{total_population}: 1861923. Total population size. \cr
\code{Sp0, Ss0, Ep0, Es0, Ip0, Is0, R0, Incidence0}: initial conditions to run the model for each compartment - these are given as prevalence and the initial conditions calculated in this function. List. Default: \code{NULL}.
If \code{NULL}: 0.1\% RSV prevalence is assumed for people during the primary infection, which is seeded at the beginning of the simulation.
All other people are assumed to be susceptible to their primary infection.
}
\description{
Provides the default parameters to run the RSV model.
}
